fixing

+++ controllers/adminController.js
@@ -1,59 +1,68 @@
 require('dotenv').config();
 const { response, request } = require('express');
-const User = require('../models/userModel');
-const Products = require('../models/cardsModel');
-const Cancha = require('../models/canchaModel')
+const { User } = require('../models/userModel');
+const { Product } = require('../models/productModel');
+const { Cancha } = require('../models/canchaModel')
 
 
 const createProduct = async (request, response) => {
-  const { Title, description, id, img, price} = request.body
+  const { Title, description, price, Url} = request.body
   try {
-    //we use the schema from the userModel
-    const NewProduct = new Products ({
-      Title,
-      id,
-      img,
-      description,
-      price,
-    });
-    //save the new product in the data base
-    await NewProduct.save();
-    response.status(200).json({message: 'se creo con exito'});
+    const product = await Product.findOne({Title, description})
+    if(product) {
+      return response.status(400).json({message: 'ya existe el producto'});
+    }
+    if(!product){
+      //we use the schema from the userModel
+      const NewProduct = new Product ({
+        Title,
+        description,
+        price,
+        Url
+      });
+      //save the new product in the data base
+      await NewProduct.save();
+      response.status(200).json({message: 'se creo con exito'});
+    }
   } catch (error) {
     response.status(400).json({message: 'no se pudo crear.'});
   }
 };
 
-const getAllProducts = async () => {
+const getAllProducts = async (request, response) => {
   try {
     //we use find({}) to get all productos from the server
-    const products = await Products.find({});
+    const products = await Product.find({});
     response.status(200).json(products);
   } catch (error) {
     response.status(500).json({ message: 'No se pudieron encontrar los productos' });
   }
 };
 
-const createCancha = async () => {
-  const { Title, description, id, img, Date} = request.body
+const createCancha = async (request, response) => {
+  const { Title, description, Url} = request.body
   try {
-    //we use the schema from the canchaModel
-    const NewCancha = new Cancha ({
-      Title,
-      id,
-      img,
-      description,
-      Date
-    });
-    //then we save the new cancha
-    await NewCancha.save();
-    response.status(200).json({message: 'se creo con exito'});
+    const cancha = await Cancha.findOne({Title, description})
+    if(cancha) {
+      return response.status(400).json({message: 'ya existe la cancha'});
+    }
+    if(!cancha) {
+      //we use the schema from the canchaModel
+      const NewCancha = new Cancha ({
+        Title,
+        Url,
+        description
+      });
+      //then we save the new cancha
+      await NewCancha.save();
+      response.status(200).json({message: 'se creo con exito'});
+    }
   } catch (error) {
     response.status(400).json({message: 'no se pudo crear.'});
   }
 };
 
-const getAllCancha = async () => {
+const getAllCancha = async (request, response) => {
   try {
     // we use find({}) to get all from canchas in the data base
     const canchas = await Cancha.find({});
@@ -63,7 +72,7 @@ const getAllCancha = async () => {
   }
 };
 
-const getAllUsers = async() => {
+const getAllUsers = async(request, response) => {
   try {
     // we use find({}) to get all from users in the data base
     const users = await User.find({});
@@ -111,8 +120,9 @@ const DeleteUser = async (request, response) => {
     if(!user){
       return response.status(400).json({message:'no se encontro el usuario'});
     }
-
-    await User.deleteOne({_id:id})
+    if(user){
+      await User.deleteOne({_id:id})
+    }
 
   } catch (error) {
     response.status(400).json({message:'no se pudo realizar la accion'})
@@ -124,13 +134,15 @@ const DeleteProducts = async (request, response) => {
   const { id } = request.body;
   try {
     // the number from the front is a id, then use these id to find the product in the data base
-    const product = await Products.findOne({_id: id});
+    const product = await Product.findOne({_id: id});
     // we check if the product exist
     if(!product){
       return response.status(400).json({message: 'no existe el producto'})
     };
     // and if it exist we delete the product by id
-    await Products.deleteOne({_id: id});
+    if(product){
+      await Product.deleteOne({_id: id});
+    }
 
   } catch (error) {
     response.status(400).json({message: 'no se pudo realizar la accion'});
@@ -148,10 +160,12 @@ const DeleteCanchas = async (request, response) => {
       return response.status(400).json({message:'no existe la cancha'})
     }
     //if it exist we delete the cancha by id
-    await Cancha.deleteOne({_id: id})
+    if(canchas){
+      await Cancha.deleteOne({_id: id})
+    }
   } catch (error) {
     response.status(400).json({message: 'no se pudo realizar la accion'})
   }
 }
 
-module.exports = { createProduct, createCancha, getAllProducts, getAllCancha, getAllUsers, DeleteCanchas, DeleteProducts, DeleteUser}
+module.exports = { createProduct, createCancha, getAllProducts, getAllCancha, getAllUsers, DeleteCanchas, DeleteProducts, DeleteUser, changeRole }
